Separate responsibilities in the translation API
