+++ betty/project/extension/trees/__init__.py
@@ -41,7 +41,7 @@ async def _generate_people_json_for_locale(
     project = event.project
     url_generator = await project.url_generator
     localizers = await project.localizers
-    localizer = await localizers.get(locale)
+    localizer = localizers.get(locale)
     private_label = localizer._("private")
     people = {
         person.id: {
