+++ betty/tests/locale/translation/project/test_extension.py
@@ -5,14 +5,11 @@ from pathlib import Path
 import pytest
 from typing_extensions import override
 
-from betty import ASSETS_DIRECTORY_PATH
 from betty.exception import UserFacingException
-from betty.locale.translation import (
+from betty.locale.translation.project.extension import (
     assert_extension_assets_directory_path,
     assert_extension_has_assets_directory_path,
-    update_dev_translations,
 )
-from betty.test_utils.locale import PotFileTestBase
 from betty.test_utils.project.extension import DummyExtension
 
 
@@ -45,21 +42,3 @@ def test_assert_extension_has_assets_directory_path__with_assets_directory() -> 
         assert_extension_has_assets_directory_path(_DummyExtensionWithAssetsDirectory)
         == _DummyExtensionWithAssetsDirectory
     )
-
-
-class TestPotFile(PotFileTestBase):
-    @override
-    def assets_directory_path(self) -> Path:
-        return ASSETS_DIRECTORY_PATH
-
-    @override
-    def command(self) -> str:
-        return "betty dev-update-translations"  # pragma: no cover
-
-    @override
-    async def update_translations(
-        self, output_assets_directory_path_override: Path
-    ) -> None:
-        await update_dev_translations(
-            _output_assets_directory_path_override=output_assets_directory_path_override
-        )
